moved converToCloudEvent function to message implementation and reorganized imports
